Add warning if SSH_ASKPASS is not set.
SSH_ASKPASS is required to be set if user is to receive a graphical
popup asking for the ssh password if necessary. This adds a check to
tell the user when this is not set in the terminal so the user can at
least know what they have to do to fix the error the might be having.