Add SingleAppendAction class
Regardless of how many times a destination is passed the input files
should only be uploaded to it once, however none of the default actions
allow appending an item only once to a list in the argparse namespace
object. This action only appends if no object of the the same type is
already in dest.